requirements.txt
xmlschema==2.5.0
pandas>=2.2.0
lxml==4.9.2
pymongo==4.5.0 
python-dotenv==1.0.0  